# perf: use string concat instead of array join when serialising cookie